убрал отдельный класс и запихнул в существующий
