+++ Binding/Gamma.Binding/TreeModels/ColumnConfig/ICustomRendererMapping.cs
@@ -1,0 +1,6 @@
+﻿namespace Gamma.ColumnConfig
+{
+	public interface ICustomRendererMapping {
+		bool Custom { get; }
+	}
+}
